:hammer: Extract SearchPasteData and QueryPasteTag interfaces from DAO layer
