+++ shared/src/commonMain/kotlin/com/crosspaste/db/paste/QueryPasteTag.kt
@@ -1,0 +1,9 @@
+package com.crosspaste.db.paste
+
+import com.crosspaste.paste.PasteTag
+import kotlinx.coroutines.flow.Flow
+
+interface QueryPasteTag {
+
+    fun getAllTagsFlow(): Flow<List<PasteTag>>
+}
